Massage to make springmvc-tomcat work, see a change in image size, but
not much improvement in RSS
Build memory: 6.66GB
Image build time: 82.5s
RSS memory: 70.9M
Image size: 57.1M
Startup time: 0.076 (JVM running for 0.078)

+++ spring-graal-native-samples/springmvc-tomcat/compile.sh
@@ -3,7 +3,8 @@
 ARTIFACT=springmvc-tomcat
 MAINCLASS=com.example.tomcat.TomcatApplication
 VERSION=0.0.1-SNAPSHOT
-FEATURE=../../../../spring-graal-native/target/spring-graal-native-0.6.0.RELEASE.jar
+FEATURE_VERSION=0.6.0.RELEASE
+FEATURE=../../../../spring-graal-native/target/spring-graal-native-$FEATURE_VERSION.jar
 
 GREEN='\033[0;32m'
 RED='\033[0;31m'
@@ -25,14 +26,26 @@ cp -R META-INF BOOT-INF/classes
 LIBPATH=`find BOOT-INF/lib | tr '\n' ':'`
 CP=BOOT-INF/classes:$LIBPATH:$FEATURE
 
-echo "Generating reflection files for $ARTIFACT"
+if [ ! -f "$FEATURE" ]; then
+    printf "${RED}FAILURE${NC}: $FEATURE does not exist, please build the root project before building this sample.\n"
+    exit 1
+fi
+
+echo "Performing class analysis on $ARTIFACT"
+CLASS_AGENT=$HOME/.m2/repository/org/springframework/experimental/spring-graal-native-feature/$FEATURE_VERSION/spring-graal-native-feature-$FEATURE_VERSION-classlist-agent.jar
 rm -rf graal/META-INF 2>/dev/null
 mkdir -p graal/META-INF/native-image
-java -agentlib:native-image-agent=config-output-dir=graal/META-INF/native-image -cp $CP $MAINCLASS >> output.txt 2>&1 &
+java -javaagent:$CLASS_AGENT -cp $CP $MAINCLASS >> output.txt 2>&1 &
+
 PID=$!
 sleep 3
 curl -m 1 http://localhost:8080 > /dev/null 2>&1
-sleep 1 && kill $PID || kill -9 $PID
+sleep 1 && kill -9 $PID
+
+cat output.txt |grep "Class\-Agent\-Transform\: " 2>&1 > class_histogram.txt
+echo "Starting classpath reduction:" >> output.txt
+java -cp $CP org.springframework.graal.util.OptimizeClassPath `pwd` class_histogram.txt $CP  >> output.txt 2>&1
+cat output.txt |grep "^Deleted\:" 2>&1 > class_deleted.txt
 
 GRAALVM_VERSION=`native-image --version`
 echo "Compiling $ARTIFACT with $GRAALVM_VERSION"
@@ -47,6 +60,7 @@ echo "Compiling $ARTIFACT with $GRAALVM_VERSION"
   -Dsun.rmi.transport.tcp.maxConnectionThreads=0 \
   -Dspring.graal.remove-unused-autoconfig=true \
   -Dspring.graal.remove-yaml-support=true \
+  -Dspring.graal.verbose=true \
   -cp $CP $MAINCLASS >> output.txt ; } 2>> output.txt
 
 if [[ -f $ARTIFACT ]]
